requirements.txt
diffusers
torch
accelerate
wandb
matplotlib
torchvision
controlnet-aux
pillow
trl
numpy